+++ CHANGELOG.md
@@ -10,13 +10,19 @@ This release changes the install directory from ~/Mycodo to /opt/Mycodo. This ne
 
  - Fix some Inputs potentially sending integers to influxdb when it should be float
  - Fix influxdb client install if server is already installed
+ - Fix Select options of web UI not properly displaying the saved value
+ - Fix note attachments not being copied during upgrade
 
 ### Features
 
  - Add Output: GPIO On/Off using pinctrl (First Pi 5-compatible Output)
+ - Add Output: PWM MQTT Publish
+ - Add Output: GP8403 2-Channel DAC (0-10 VDC) ([#1354](https://github.com/kizniche/Mycodo/issues/1354))
+ - Add API Endpoint: /notes/create to create a Note ([#1357](https://github.com/kizniche/Mycodo/issues/1357))
  - Add ability to switch displaying hostname with custom text
  - Add Step Line Series Type to Graph (Synchronous) Widget
  - Add controller_restart as client endpoint
+ - Add option for custom CSS
 
 ### Miscellaneous